backend/requirements.txt
fastapi
uvicorn
langchain
huggingface_hub
langchain_core
python-dotenv

